requirements.txt
accelerate==1.0.0
bitsandbytes==0.44.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
filelock==3.13.1
Flask==3.0.3
fsspec==2024.2.0
huggingface-hub==0.25.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.3
packaging==24.1
pillow==10.2.0
psutil==6.0.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
setuptools==70.0.0
sympy==1.13.1
tokenizers==0.20.1
torch==2.5.0+cu121
torchaudio==2.5.0+cu121
torchvision==0.20.0+cu121
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.9.0
urllib3==2.2.3
Werkzeug==3.0.4
